fix: check for empty configuration file at startup and overwrite it if necessary.

+++ core/src/gaiasky/desktop/GaiaSkyDesktop.java
@@ -44,7 +44,6 @@ import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.util.Map;
 
@@ -63,7 +62,9 @@ public class GaiaSkyDesktop implements IObserver {
     private static final int MIN_GLSL_MINOR = 3;
     private static final String MIN_GLSL = MIN_GLSL_MAJOR + "." + MIN_GLSL_MINOR;
     private static boolean REST_ENABLED = false;
-    /** Running with an unsupported Java version. **/
+    /**
+     * Running with an unsupported Java version.
+     **/
     private static boolean JAVA_VERSION_PROBLEM_FLAG = false;
     private static CLIArgs cliArgs;
     /**
@@ -238,7 +239,6 @@ public class GaiaSkyDesktop implements IObserver {
      *
      * @return True if the configuration file has been initialized or
      * overwritten with the default one, false otherwise.
-     *
      * @throws IOException If the file fails to be written successfully.
      */
     private static boolean initConfigFile(final boolean vr) throws IOException {
@@ -261,11 +261,15 @@ public class GaiaSkyDesktop implements IObserver {
             }
 
             // Check latest version
-            if (!userProps.containsKey("configVersion")) {
-                out.println("Properties file version not found, overwriting with new version (" + internalVersion + ")");
+            if (userProps == null) {
+                out.println("Your current configuration file is corrupted! Overwriting...");
+                userConfExists = false;
+                overwrite = true;
+            } else if (!userProps.containsKey("configVersion")) {
+                out.println("Configuration file version not found, overwriting with new version (" + internalVersion + ")");
                 overwrite = true;
             } else if ((Integer) userProps.get("configVersion") < internalVersion) {
-                out.println("Properties file version mismatch, overwriting with new version: found " + userProps.get("version") + ", required " + internalVersion);
+                out.println("Configuration file version mismatch, overwriting with new version: found " + userProps.get("version") + ", required " + internalVersion);
                 overwrite = true;
             }
         } else {
@@ -273,7 +277,7 @@ public class GaiaSkyDesktop implements IObserver {
             try {
                 Path propertiesFile = SysUtils.getConfigDir().resolve(vr ? "global.vr.properties" : "global.properties");
                 if (Files.exists(propertiesFile)) {
-                    out.println("Old properties file detected!");
+                    out.println("Old configuration file detected!");
                     out.println("    -> Converting " + propertiesFile + " to " + userFolderConfFile);
                     SettingsMorph.morphSettings(propertiesFile, userFolderConfFile);
                     // Move old properties file so that they are not converted on the next run
@@ -291,8 +295,8 @@ public class GaiaSkyDesktop implements IObserver {
         if (overwrite || !userConfExists) {
             // Copy file
             if (Files.exists(confFolder) && Files.isDirectory(confFolder)) {
-                // Back up user configuration, if it exists.
-                if (Files.exists(userFolderConfFile)) {
+                // Back up user configuration, if it exists and contains data.
+                if (Files.exists(userFolderConfFile) && userFolderConfFile.toFile().length() > 0) {
                     Path backup = userFolderConfFile.getParent().resolve(userFolderConfFile.getFileName() + "." + LocalDateTime.now().toString().replaceAll("[^a-zA-Z0-9_\\.\\-]", "_"));
                     GlobalResources.copyFile(userFolderConfFile, backup, true);
                 }
@@ -615,31 +619,31 @@ public class GaiaSkyDesktop implements IObserver {
     @Override
     public void notify(final Event event, Object source, final Object... data) {
         switch (event) {
-        case SCENE_LOADED -> {
-            if (REST_ENABLED) {
-                /*
-                 * Notify REST server that GUI is loaded and everything should be in a
-                 * well-defined state
-                 */
-                try {
-                    RESTServer.activate();
-                } catch (SecurityException | IllegalArgumentException e) {
-                    logger.error(e);
+            case SCENE_LOADED -> {
+                if (REST_ENABLED) {
+                    /*
+                     * Notify REST server that GUI is loaded and everything should be in a
+                     * well-defined state
+                     */
+                    try {
+                        RESTServer.activate();
+                    } catch (SecurityException | IllegalArgumentException e) {
+                        logger.error(e);
+                    }
                 }
             }
-        }
-        case DISPOSE -> {
-            if (REST_ENABLED) {
-                /* Shutdown REST server thread on termination */
-                try {
-                    RESTServer.dispose();
-                } catch (SecurityException | IllegalArgumentException e) {
-                    logger.error(e);
+            case DISPOSE -> {
+                if (REST_ENABLED) {
+                    /* Shutdown REST server thread on termination */
+                    try {
+                        RESTServer.dispose();
+                    } catch (SecurityException | IllegalArgumentException e) {
+                        logger.error(e);
+                    }
                 }
             }
-        }
-        default -> {
-        }
+            default -> {
+            }
         }
     }
 
@@ -647,30 +651,43 @@ public class GaiaSkyDesktop implements IObserver {
      * Program CLI arguments.
      */
     private static class CLIArgs {
-        @Parameter(names = { "-h", "--help" }, description = "Show program options and usage information.", help = true, order = 0) private boolean help = false;
+        @Parameter(names = {"-h", "--help"}, description = "Show program options and usage information.", help = true, order = 0)
+        private boolean help = false;
 
-        @Parameter(names = { "-v", "--version" }, description = "List Gaia Sky version and relevant information.", order = 1) private boolean version = false;
+        @Parameter(names = {"-v", "--version"}, description = "List Gaia Sky version and relevant information.", order = 1)
+        private boolean version = false;
 
-        @Parameter(names = { "-i", "--asciiart" }, description = "Add nice ascii art to --version information.", order = 1) private boolean asciiArt = false;
+        @Parameter(names = {"-i", "--asciiart"}, description = "Add nice ascii art to --version information.", order = 1)
+        private boolean asciiArt = false;
 
-        @Parameter(names = { "-s", "--skip-welcome" }, description = "Skip the welcome screen if possible (base-data package must be present).", order = 2) private boolean skipWelcome = false;
+        @Parameter(names = {"-s", "--skip-welcome"}, description = "Skip the welcome screen if possible (base-data package must be present).", order = 2)
+        private boolean skipWelcome = false;
 
-        @Parameter(names = { "-p", "--properties" }, description = "Specify the location of the properties file.", order = 4) private String propertiesFile = null;
+        @Parameter(names = {"-p", "--properties"}, description = "Specify the location of the properties file.", order = 4)
+        private String propertiesFile = null;
 
-        @Parameter(names = { "-a", "--assets" }, description = "Specify the location of the assets folder. If not present, the default assets location (in the installation folder) is used.", order = 5) private String assetsLocation = null;
+        @Parameter(names = {"-a", "--assets"}, description = "Specify the location of the assets folder. If not present, the default assets location (in the installation folder) is used.", order = 5)
+        private String assetsLocation = null;
 
-        @Parameter(names = { "-vr", "--openvr" }, description = "Launch in Virtual Reality mode. Gaia Sky will attempt to create a VR context through OpenVR.", order = 6) private boolean vr = false;
+        @Parameter(names = {"-vr", "--openvr"}, description = "Launch in Virtual Reality mode. Gaia Sky will attempt to create a VR context through OpenVR.", order = 6)
+        private boolean vr = false;
 
-        @Parameter(names = { "-e", "--externalview" }, description = "Create a window with a view of the scene and no UI.", order = 7) private boolean externalView = false;
+        @Parameter(names = {"-e", "--externalview"}, description = "Create a window with a view of the scene and no UI.", order = 7)
+        private boolean externalView = false;
 
-        @Parameter(names = { "-n", "--noscript" }, description = "Do not start the scripting server. Useful to run more than one Gaia Sky instance at once in the same machine.", order = 8) private boolean noScriptingServer = false;
+        @Parameter(names = {"-n", "--noscript"}, description = "Do not start the scripting server. Useful to run more than one Gaia Sky instance at once in the same machine.", order = 8)
+        private boolean noScriptingServer = false;
 
-        @Parameter(names = { "-d", "--debug" }, description = "Launch in debug mode. Prints out debug information from Gaia Sky to the logs.", order = 9) private boolean debug = false;
+        @Parameter(names = {"-d", "--debug"}, description = "Launch in debug mode. Prints out debug information from Gaia Sky to the logs.", order = 9)
+        private boolean debug = false;
 
-        @Parameter(names = { "-g", "--gpudebug" }, description = "Activate OpenGL debug mode. Prints out debug information from OpenGL to the standard output.", order = 10) private boolean debugGpu = false;
+        @Parameter(names = {"-g", "--gpudebug"}, description = "Activate OpenGL debug mode. Prints out debug information from OpenGL to the standard output.", order = 10)
+        private boolean debugGpu = false;
 
-        @Parameter(names = { "-l", "--headless" }, description = "Use headless (windowless) mode, for servers.", order = 11) private boolean headless = false;
+        @Parameter(names = {"-l", "--headless"}, description = "Use headless (windowless) mode, for servers.", order = 11)
+        private boolean headless = false;
 
-        @Parameter(names = { "--safemode" }, description = "Activate safe graphics mode. This forces the creation of an OpenGL 3.2 context, and disables float buffers and tessellation.", order = 12) private boolean safeMode = false;
+        @Parameter(names = {"--safemode"}, description = "Activate safe graphics mode. This forces the creation of an OpenGL 3.2 context, and disables float buffers and tessellation.", order = 12)
+        private boolean safeMode = false;
     }
 }
